// ChoicePillButtons/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    value: ComponentFramework.PropertyTypes.OptionSetProperty;
    color1: ComponentFramework.PropertyTypes.StringProperty;
    color2: ComponentFramework.PropertyTypes.StringProperty;
    color3: ComponentFramework.PropertyTypes.StringProperty;
    color4: ComponentFramework.PropertyTypes.StringProperty;
    color5: ComponentFramework.PropertyTypes.StringProperty;
    value1: ComponentFramework.PropertyTypes.WholeNumberProperty;
    value2: ComponentFramework.PropertyTypes.WholeNumberProperty;
    value3: ComponentFramework.PropertyTypes.WholeNumberProperty;
    value4: ComponentFramework.PropertyTypes.WholeNumberProperty;
    value5: ComponentFramework.PropertyTypes.WholeNumberProperty;
}
export interface IOutputs {
    value?: number;
}